Created Colors Grid and Modified Top bar
The top bar was modified to display itself at less than 100% of the width of the page.

The grid where the user is going to interact with the colors has been created and styled correctly. The creation of it was needed to give the user a place to follow the game visually.

Many other styles were added and created with SASS