Created Starting Functions, Modified HTML and CSS
The starting functions that are always active during the standby of the game have been created.
One blocks the input of the user so he cant produce any input while the game is running outside of his turn.
The second and the third ones update respectively the status bar according to how the state of the game is going.

The HTML has been modified to be able to properly call the elements with their class names. The status bar elements have also been modified to properly display them with JS.

CSS has been modified by adjusting some margins on the status bar and also the transition property has been added for when its going to be used while changing opacity.