Created Function Unblock User Input and Handle User Input
The functions that allow the user to produce inputs on the game interface has been created.
Also, the function that highlights which box the user selects has also been created.
The time that the user has to play has also been inserted. A simple calculation that involves the quantity of boxes the machine has in it's sequence plus one, multiplied by 1000 ms to provide enough time to the user for playing.