Modified HTML elements ID and added comparation between sequences
The if statement that compares if the user sequence is equal to the machine sequence has been added.
The user sequence reset has also moved out of the handling of the user input to not reset every time it selects a new color on the same round.

We have also given the color boxes a specific ID to compare them with our JavaScript code.