Modified Readme and index.js, added Images and pseudocode file
The readme was edited to display a detailed description of how the project needs to be run, was built and other miscellaneous info. The index.js file was modified by removing the pseudocode to its own independant file. The images that are going to be used in the Readme where also added to the project